supabase実装

+++ db/migrate/20241224000001_remove_font_size_from_users.rb
@@ -1,5 +1,7 @@
 class RemoveFontSizeFromUsers < ActiveRecord::Migration[7.0]
   def change
-    remove_column :users, :font_size, :string
+    if column_exists?(:users, :font_size)
+      remove_column :users, :font_size, :string
+    end
   end
-end 
+end
